[revm] Instructions sorted
