+++ bins/revme/Cargo.toml
@@ -16,7 +16,7 @@ hashbrown = "0.12"
 hex = "0.4"
 indicatif = "0.16"
 plain_hasher = "0.2"
-primitive-types = { version = "0.10", features = ["rlp", "serde"] }
+primitive-types = { version = "0.11", features = ["rlp", "serde"] }
 
 revm = { path = "../../crates/revm", version = "1.1", default-features = false, features = ["web3db","std","k256"] }
 rlp = { version = "0.5", default-features = false }
