+++ crates/revm/src/instructions.rs
@@ -2,9 +2,13 @@
 mod macros;
 mod arithmetic;
 mod bitwise;
+mod control;
+mod host;
+mod host_env;
 mod i256;
-mod misc;
+mod memory;
 pub mod opcode;
+mod stack;
 mod system;
 
 pub use opcode::{OpCode, OPCODE_JUMPMAP};
@@ -120,123 +124,123 @@ pub fn eval<H: Host, S: Spec>(opcode: u8, machine: &mut Interpreter, host: &mut 
         opcode::SHA3 => system::sha3(machine),
 
         opcode::ADDRESS => system::address(machine),
-        opcode::BALANCE => system::balance::<H, S>(machine, host),
-        opcode::SELFBALANCE => system::selfbalance::<H, S>(machine, host),
-        opcode::CODESIZE => misc::codesize(machine),
-        opcode::CODECOPY => misc::codecopy(machine),
-        opcode::CALLDATALOAD => misc::calldataload(machine),
-        opcode::CALLDATASIZE => misc::calldatasize(machine),
-        opcode::CALLDATACOPY => misc::calldatacopy(machine),
-        opcode::POP => misc::pop(machine),
-        opcode::MLOAD => misc::mload(machine),
-        opcode::MSTORE => misc::mstore(machine),
-        opcode::MSTORE8 => misc::mstore8(machine),
-        opcode::JUMP => misc::jump(machine),
-        opcode::JUMPI => misc::jumpi(machine),
-        opcode::PC => misc::pc(machine),
-        opcode::MSIZE => misc::msize(machine),
-        opcode::JUMPDEST => misc::jumpdest(machine),
-        opcode::PUSH1 => misc::push::<1>(machine),
-        opcode::PUSH2 => misc::push::<2>(machine),
-        opcode::PUSH3 => misc::push::<3>(machine),
-        opcode::PUSH4 => misc::push::<4>(machine),
-        opcode::PUSH5 => misc::push::<5>(machine),
-        opcode::PUSH6 => misc::push::<6>(machine),
-        opcode::PUSH7 => misc::push::<7>(machine),
-        opcode::PUSH8 => misc::push::<8>(machine),
-        opcode::PUSH9 => misc::push::<9>(machine),
-        opcode::PUSH10 => misc::push::<10>(machine),
-        opcode::PUSH11 => misc::push::<11>(machine),
-        opcode::PUSH12 => misc::push::<12>(machine),
-        opcode::PUSH13 => misc::push::<13>(machine),
-        opcode::PUSH14 => misc::push::<14>(machine),
-        opcode::PUSH15 => misc::push::<15>(machine),
-        opcode::PUSH16 => misc::push::<16>(machine),
-        opcode::PUSH17 => misc::push::<17>(machine),
-        opcode::PUSH18 => misc::push::<18>(machine),
-        opcode::PUSH19 => misc::push::<19>(machine),
-        opcode::PUSH20 => misc::push::<20>(machine),
-        opcode::PUSH21 => misc::push::<21>(machine),
-        opcode::PUSH22 => misc::push::<22>(machine),
-        opcode::PUSH23 => misc::push::<23>(machine),
-        opcode::PUSH24 => misc::push::<24>(machine),
-        opcode::PUSH25 => misc::push::<25>(machine),
-        opcode::PUSH26 => misc::push::<26>(machine),
-        opcode::PUSH27 => misc::push::<27>(machine),
-        opcode::PUSH28 => misc::push::<28>(machine),
-        opcode::PUSH29 => misc::push::<29>(machine),
-        opcode::PUSH30 => misc::push::<30>(machine),
-        opcode::PUSH31 => misc::push::<31>(machine),
-        opcode::PUSH32 => misc::push::<32>(machine),
-        opcode::DUP1 => misc::dup::<1>(machine),
-        opcode::DUP2 => misc::dup::<2>(machine),
-        opcode::DUP3 => misc::dup::<3>(machine),
-        opcode::DUP4 => misc::dup::<4>(machine),
-        opcode::DUP5 => misc::dup::<5>(machine),
-        opcode::DUP6 => misc::dup::<6>(machine),
-        opcode::DUP7 => misc::dup::<7>(machine),
-        opcode::DUP8 => misc::dup::<8>(machine),
-        opcode::DUP9 => misc::dup::<9>(machine),
-        opcode::DUP10 => misc::dup::<10>(machine),
-        opcode::DUP11 => misc::dup::<11>(machine),
-        opcode::DUP12 => misc::dup::<12>(machine),
-        opcode::DUP13 => misc::dup::<13>(machine),
-        opcode::DUP14 => misc::dup::<14>(machine),
-        opcode::DUP15 => misc::dup::<15>(machine),
-        opcode::DUP16 => misc::dup::<16>(machine),
+        opcode::BALANCE => host::balance::<H, S>(machine, host),
+        opcode::SELFBALANCE => host::selfbalance::<H, S>(machine, host),
+        opcode::CODESIZE => system::codesize(machine),
+        opcode::CODECOPY => system::codecopy(machine),
+        opcode::CALLDATALOAD => system::calldataload(machine),
+        opcode::CALLDATASIZE => system::calldatasize(machine),
+        opcode::CALLDATACOPY => system::calldatacopy(machine),
+        opcode::POP => stack::pop(machine),
+        opcode::MLOAD => memory::mload(machine),
+        opcode::MSTORE => memory::mstore(machine),
+        opcode::MSTORE8 => memory::mstore8(machine),
+        opcode::JUMP => control::jump(machine),
+        opcode::JUMPI => control::jumpi(machine),
+        opcode::PC => control::pc(machine),
+        opcode::MSIZE => memory::msize(machine),
+        opcode::JUMPDEST => control::jumpdest(machine),
+        opcode::PUSH1 => stack::push::<1>(machine),
+        opcode::PUSH2 => stack::push::<2>(machine),
+        opcode::PUSH3 => stack::push::<3>(machine),
+        opcode::PUSH4 => stack::push::<4>(machine),
+        opcode::PUSH5 => stack::push::<5>(machine),
+        opcode::PUSH6 => stack::push::<6>(machine),
+        opcode::PUSH7 => stack::push::<7>(machine),
+        opcode::PUSH8 => stack::push::<8>(machine),
+        opcode::PUSH9 => stack::push::<9>(machine),
+        opcode::PUSH10 => stack::push::<10>(machine),
+        opcode::PUSH11 => stack::push::<11>(machine),
+        opcode::PUSH12 => stack::push::<12>(machine),
+        opcode::PUSH13 => stack::push::<13>(machine),
+        opcode::PUSH14 => stack::push::<14>(machine),
+        opcode::PUSH15 => stack::push::<15>(machine),
+        opcode::PUSH16 => stack::push::<16>(machine),
+        opcode::PUSH17 => stack::push::<17>(machine),
+        opcode::PUSH18 => stack::push::<18>(machine),
+        opcode::PUSH19 => stack::push::<19>(machine),
+        opcode::PUSH20 => stack::push::<20>(machine),
+        opcode::PUSH21 => stack::push::<21>(machine),
+        opcode::PUSH22 => stack::push::<22>(machine),
+        opcode::PUSH23 => stack::push::<23>(machine),
+        opcode::PUSH24 => stack::push::<24>(machine),
+        opcode::PUSH25 => stack::push::<25>(machine),
+        opcode::PUSH26 => stack::push::<26>(machine),
+        opcode::PUSH27 => stack::push::<27>(machine),
+        opcode::PUSH28 => stack::push::<28>(machine),
+        opcode::PUSH29 => stack::push::<29>(machine),
+        opcode::PUSH30 => stack::push::<30>(machine),
+        opcode::PUSH31 => stack::push::<31>(machine),
+        opcode::PUSH32 => stack::push::<32>(machine),
+        opcode::DUP1 => stack::dup::<1>(machine),
+        opcode::DUP2 => stack::dup::<2>(machine),
+        opcode::DUP3 => stack::dup::<3>(machine),
+        opcode::DUP4 => stack::dup::<4>(machine),
+        opcode::DUP5 => stack::dup::<5>(machine),
+        opcode::DUP6 => stack::dup::<6>(machine),
+        opcode::DUP7 => stack::dup::<7>(machine),
+        opcode::DUP8 => stack::dup::<8>(machine),
+        opcode::DUP9 => stack::dup::<9>(machine),
+        opcode::DUP10 => stack::dup::<10>(machine),
+        opcode::DUP11 => stack::dup::<11>(machine),
+        opcode::DUP12 => stack::dup::<12>(machine),
+        opcode::DUP13 => stack::dup::<13>(machine),
+        opcode::DUP14 => stack::dup::<14>(machine),
+        opcode::DUP15 => stack::dup::<15>(machine),
+        opcode::DUP16 => stack::dup::<16>(machine),
 
-        opcode::SWAP1 => misc::swap::<1>(machine),
-        opcode::SWAP2 => misc::swap::<2>(machine),
-        opcode::SWAP3 => misc::swap::<3>(machine),
-        opcode::SWAP4 => misc::swap::<4>(machine),
-        opcode::SWAP5 => misc::swap::<5>(machine),
-        opcode::SWAP6 => misc::swap::<6>(machine),
-        opcode::SWAP7 => misc::swap::<7>(machine),
-        opcode::SWAP8 => misc::swap::<8>(machine),
-        opcode::SWAP9 => misc::swap::<9>(machine),
-        opcode::SWAP10 => misc::swap::<10>(machine),
-        opcode::SWAP11 => misc::swap::<11>(machine),
-        opcode::SWAP12 => misc::swap::<12>(machine),
-        opcode::SWAP13 => misc::swap::<13>(machine),
-        opcode::SWAP14 => misc::swap::<14>(machine),
-        opcode::SWAP15 => misc::swap::<15>(machine),
-        opcode::SWAP16 => misc::swap::<16>(machine),
+        opcode::SWAP1 => stack::swap::<1>(machine),
+        opcode::SWAP2 => stack::swap::<2>(machine),
+        opcode::SWAP3 => stack::swap::<3>(machine),
+        opcode::SWAP4 => stack::swap::<4>(machine),
+        opcode::SWAP5 => stack::swap::<5>(machine),
+        opcode::SWAP6 => stack::swap::<6>(machine),
+        opcode::SWAP7 => stack::swap::<7>(machine),
+        opcode::SWAP8 => stack::swap::<8>(machine),
+        opcode::SWAP9 => stack::swap::<9>(machine),
+        opcode::SWAP10 => stack::swap::<10>(machine),
+        opcode::SWAP11 => stack::swap::<11>(machine),
+        opcode::SWAP12 => stack::swap::<12>(machine),
+        opcode::SWAP13 => stack::swap::<13>(machine),
+        opcode::SWAP14 => stack::swap::<14>(machine),
+        opcode::SWAP15 => stack::swap::<15>(machine),
+        opcode::SWAP16 => stack::swap::<16>(machine),
 
-        opcode::RETURN => misc::ret(machine),
-        opcode::REVERT => misc::revert::<S>(machine),
+        opcode::RETURN => control::ret(machine),
+        opcode::REVERT => control::revert::<S>(machine),
         opcode::INVALID => Return::InvalidOpcode,
-        opcode::BASEFEE => system::basefee::<H, S>(machine, host),
-        opcode::ORIGIN => system::origin(machine, host),
+        opcode::BASEFEE => host_env::basefee::<H, S>(machine, host),
+        opcode::ORIGIN => host_env::origin(machine, host),
         opcode::CALLER => system::caller(machine),
         opcode::CALLVALUE => system::callvalue(machine),
-        opcode::GASPRICE => system::gasprice(machine, host),
-        opcode::EXTCODESIZE => system::extcodesize::<H, S>(machine, host),
-        opcode::EXTCODEHASH => system::extcodehash::<H, S>(machine, host),
-        opcode::EXTCODECOPY => system::extcodecopy::<H, S>(machine, host),
+        opcode::GASPRICE => host_env::gasprice(machine, host),
+        opcode::EXTCODESIZE => host::extcodesize::<H, S>(machine, host),
+        opcode::EXTCODEHASH => host::extcodehash::<H, S>(machine, host),
+        opcode::EXTCODECOPY => host::extcodecopy::<H, S>(machine, host),
         opcode::RETURNDATASIZE => system::returndatasize::<S>(machine),
         opcode::RETURNDATACOPY => system::returndatacopy::<S>(machine),
-        opcode::BLOCKHASH => system::blockhash(machine, host),
-        opcode::COINBASE => system::coinbase(machine, host),
-        opcode::TIMESTAMP => system::timestamp(machine, host),
-        opcode::NUMBER => system::number(machine, host),
-        opcode::DIFFICULTY => system::difficulty(machine, host),
-        opcode::GASLIMIT => system::gaslimit(machine, host),
-        opcode::SLOAD => system::sload::<H, S>(machine, host),
-        opcode::SSTORE => system::sstore::<H, S>(machine, host),
+        opcode::BLOCKHASH => host::blockhash(machine, host),
+        opcode::COINBASE => host_env::coinbase(machine, host),
+        opcode::TIMESTAMP => host_env::timestamp(machine, host),
+        opcode::NUMBER => host_env::number(machine, host),
+        opcode::DIFFICULTY => host_env::difficulty(machine, host),
+        opcode::GASLIMIT => host_env::gaslimit(machine, host),
+        opcode::SLOAD => host::sload::<H, S>(machine, host),
+        opcode::SSTORE => host::sstore::<H, S>(machine, host),
         opcode::GAS => system::gas(machine),
-        opcode::LOG0 => system::log::<H, S>(machine, 0, host),
-        opcode::LOG1 => system::log::<H, S>(machine, 1, host),
-        opcode::LOG2 => system::log::<H, S>(machine, 2, host),
-        opcode::LOG3 => system::log::<H, S>(machine, 3, host),
-        opcode::LOG4 => system::log::<H, S>(machine, 4, host),
-        opcode::SELFDESTRUCT => system::selfdestruct::<H, S>(machine, host),
-        opcode::CREATE => system::create::<H, S>(machine, false, host), //check
-        opcode::CREATE2 => system::create::<H, S>(machine, true, host), //check
-        opcode::CALL => system::call::<H, S>(machine, CallScheme::Call, host), //check
-        opcode::CALLCODE => system::call::<H, S>(machine, CallScheme::CallCode, host), //check
-        opcode::DELEGATECALL => system::call::<H, S>(machine, CallScheme::DelegateCall, host), //check
-        opcode::STATICCALL => system::call::<H, S>(machine, CallScheme::StaticCall, host), //check
-        opcode::CHAINID => system::chainid::<H, S>(machine, host),
+        opcode::LOG0 => host::log::<H, S>(machine, 0, host),
+        opcode::LOG1 => host::log::<H, S>(machine, 1, host),
+        opcode::LOG2 => host::log::<H, S>(machine, 2, host),
+        opcode::LOG3 => host::log::<H, S>(machine, 3, host),
+        opcode::LOG4 => host::log::<H, S>(machine, 4, host),
+        opcode::SELFDESTRUCT => host::selfdestruct::<H, S>(machine, host),
+        opcode::CREATE => host::create::<H, S>(machine, false, host), //check
+        opcode::CREATE2 => host::create::<H, S>(machine, true, host), //check
+        opcode::CALL => host::call::<H, S>(machine, CallScheme::Call, host), //check
+        opcode::CALLCODE => host::call::<H, S>(machine, CallScheme::CallCode, host), //check
+        opcode::DELEGATECALL => host::call::<H, S>(machine, CallScheme::DelegateCall, host), //check
+        opcode::STATICCALL => host::call::<H, S>(machine, CallScheme::StaticCall, host), //check
+        opcode::CHAINID => host_env::chainid::<H, S>(machine, host),
         _ => Return::OpcodeNotFound,
     }
 }
