+++ crates/revm_precompiles/Cargo.toml
@@ -13,7 +13,7 @@ bn = { package = "substrate-bn", version = "0.6", default-features = false }
 bytes = { version = "1.1", default-features = false }
 k256 = { version = "0.10.1", default-features = false, features = ["ecdsa", "keccak256"], optional = true }
 num = { version = "0.4.0", default-features = false, features = ["alloc"] }
-primitive-types = { version = "0.10", default-features = false, features = ["rlp"] }
+primitive-types = { version = "0.11", default-features = false, features = ["rlp"] }
 ripemd = { version = "0.1", default-features = false }
 secp256k1 = { version = "0.21.2", default-features = false, features = ["alloc", "recovery"], optional = true }
 sha2 = { version = "0.10.1", default-features = false }
